Finished arrays episode

+++ 04_complex_structures/code/exercise_4b.go
@@ -2,14 +2,14 @@ package main
 
 func main() {
 	average := average(12, 13, 14, 15, 16, 17, 19)
-	println(int(average))
+	println(average)
 }
-func average(numbers ...int) float64 {
-	sum := 0
+func average(numbers ...float64) float64 {
+	sum := 0.0
 
 	for _, number := range numbers {
 		sum += number
 	}
 
-	return float64(sum) / float64(len(numbers))
+	return sum / float64(len(numbers))
 }
